first round of reformatting
